Separated grader report from semester report
- Grader report is the source file and can be reused for multiple different levels.
- A semester report is the output file and may have different format and layout for each level.

Hence, both items are turned into separate modular components.
