+++ processor/semester_report.py
@@ -1,6 +1,3 @@
-import pandas as pd
-from termcolor import colored
-
 from docx import Document
 from docx.shared import Cm, Pt
 from docx.enum.table import WD_TABLE_ALIGNMENT
@@ -8,6 +5,7 @@ from docx.enum.table import WD_TABLE_ALIGNMENT
 import config
 import processor.helper.document as document_helper
 import processor.helper.comment_generator as cgen
+from processor.grader_report import GraderReport
 
 class Generator:
     """
@@ -16,12 +14,11 @@ class Generator:
     This class is responsible for generating a standardized DOCX file from a templated XLSX grader report.
     """
 
-    def __init__(self, grader_report_path, output_path):
+    def __init__(self, output_path, grader_report: GraderReport):
         """
         Initialize the generator instance.
 
         Args:
-            grader_report_path (str): The path of the grader report.
             output_path (str): The path of the output file.
 
         Returns:
@@ -29,173 +26,14 @@ class Generator:
         """
 
         print("[  ] Initializing generator...")
-
-        self.grader_report_path = grader_report_path
         self.output_path = output_path
-
-        self.course_info = pd.read_excel(self.grader_report_path, sheet_name = "Course Information", index_col = 0, header = 0, nrows = 7)
-        self.students = pd.read_excel(self.grader_report_path, sheet_name = "Student List", index_col = 0, header = 0, usecols = "A:C")
-        self.data_sna = pd.read_excel(self.grader_report_path, sheet_name = "Skills and Assessment", index_col = 0, header = 0)
-        self.data_pd = pd.read_excel(self.grader_report_path, sheet_name = "Personal Development", index_col = 0, header = 0)
-        self.data_final_grades = pd.read_excel(self.grader_report_path, sheet_name = "Final Grades", index_col = 0, header = 0, usecols = "A,H:I")
-        self.data_comment_mapping = pd.read_excel(self.grader_report_path, sheet_name = "Comment Mapping", index_col = 0, header = 0)
-        
-        self.__prepare_data()
-        self.__data_valid = self.validate()
-        if not self.__data_valid:
-            print(colored("\n"
-                f"Attention: There are missing values in the grader report! Please check the grader report again and make sure all data is filled.",
-                "white", "on_red"), "\n")
-
+        self.grader_report = grader_report
         print("[OK] Report generator initialized!")
 
-    def get_course_info(self, item):
-        return self.course_info.loc[item, "Value"]
-    
-    def get_student_info(self, student, item):
-        return self.students.loc[student, item]
-    
-    def get_grade_sna(self, student, assessment, raw = False):
-        data = self.data_sna.loc[student, assessment]
-
-        if raw:
-            return data
-        
-        if data == "X":
-            print(colored(f"Warning: {student} has no grade for {assessment}! Please check the grader report.", "red"))
-            return ""
-        else:
-            return data
-    
-    def get_grade_pd(self, student, item, raw = False):
-        data = self.data_pd.loc[student, item]
-
-        if raw:
-            return data
-        
-        if data == 0:
-            print(colored(
-                f"Warning: {student} has no grade for {item}! Please check the grader report. "
-                "Due to this, the grade will be set to NI to prevent software errors.", 
-                "red"))
-            return 1
-        else:
-            return data
-    
-    def get_final_grade(self, student, format, raw = False):
-        data = self.data_final_grades.loc[student, format]
-
-        if raw:
-            return data
-
-        if data == 0:
-            if format == "Final Score":
-                print(colored(
-                    f"Warning: {student} has no final score! Please check the grader report. "
-                    "Due to this, the final score will be left blank to prevent software errors.", 
-                    "red"
-                ))
-            elif format == "Letter Grade":
-                print(colored(
-                    f"Warning: {student} has no letter grade! Please check the grader report. "
-                    "Due to this, the letter grade will be left blank to X to prevent software errors.", 
-                    "red"
-                ))
-            data = str(" ")
-        
-        return data
-    
-    def __prepare_data(self):
-        # Remove NaN rows
-        self.students.dropna(inplace = True)
-
-        # Fill NaN values with default values
-        self.course_info.fillna("", inplace = True)
-        self.data_final_grades.fillna(0, inplace = True)
-        self.data_pd.fillna(0, inplace = True)
-        self.data_sna.fillna("X", inplace = True)
-
-        # Data type conversion
-        self.course_info = self.course_info.astype(str)
-        self.students = self.students.astype(str)
-        self.data_pd = self.data_pd.astype(int)
-        self.data_sna = self.data_sna.astype(str)
-        self.data_final_grades["Final Score"] = self.data_final_grades["Final Score"].round(0).astype(int)
-        self.data_comment_mapping = self.data_comment_mapping.astype(str)
-
-        # Strip whitespace from index
-        self.course_info.index = self.course_info.index.str.strip()
-        self.students.index = self.students.index.str.strip()
-        self.data_sna.index = self.data_sna.index.str.strip()
-        self.data_pd.index = self.data_pd.index.str.strip()
-        self.data_final_grades.index = self.data_final_grades.index.str.strip()
-        self.data_comment_mapping.index = self.data_comment_mapping.index.str.strip()
-
-        # Strip whitespace from columns
-        self.course_info.columns = self.course_info.columns.str.strip()
-        self.students.columns = self.students.columns.str.strip()
-        self.data_sna.columns = self.data_sna.columns.str.strip()
-        self.data_pd.columns = self.data_pd.columns.str.strip()
-        self.data_final_grades.columns = self.data_final_grades.columns.str.strip()
-        self.data_comment_mapping.columns = self.data_comment_mapping.columns.str.strip()
-
-        # Remove unnecessary columns
-        unwanted_columns = ["Normalized Grade", "Student Final Grade", "Sanity Check", "Add item…"]
-        unwanted_columns += [column for column in self.data_sna.columns if column.startswith("Unnamed")]
-        self.data_sna = self.data_sna.drop(columns = unwanted_columns)
-
-    def validate(self):
-        print("[  ] Validating data...")
-        valid = True
-        count = 0
-
-        print(f"Course Info:\n{self.course_info}\n")
-        print(f"Student List:\n{self.students}\n")
-
-        # Check if all course information is filled
-        for item in self.course_info.index:
-            if str(self.get_course_info(item)) is "": # Check if the value is empty
-                count += 1
-                valid = False
-                print(colored(f"Warning [{count}]: Course information '{item}' is not filled! Please check the grader report.", "red"))
-
-        # Check if all students have a final grade
-        for student in self.students.index:
-            if self.get_final_grade(student, "Final Score", raw = True) == 0:
-                count += 1
-                valid = False
-                print(colored(f"Warning [{count}]: {student} has no final score! Please check the grader report.", "red"))
-
-        # Check if all students have a letter grade
-        for student in self.students.index:
-            if self.get_final_grade(student, "Letter Grade", raw = True) == 0:
-                count += 1
-                valid = False
-                print(colored(f"Warning [{count}]: {student} has no letter grade! Please check the grader report.", "red"))
-
-        # Check if all students have a grade for each SNA
-        for student in self.students.index:
-            for assessment in self.data_sna.columns:
-                if self.get_grade_sna(student, assessment, raw = True) == "X":
-                    count += 1
-                    valid = False
-                    print(colored(f"Warning [{count}]: {student} has no grade for goal '{assessment}'! Please check the grader report.", "red"))
-
-        # Check if all students have a grade for each PD item
-        for student in self.students.index:
-            for item in self.data_pd.columns:
-                if self.get_grade_pd(student, item, raw = True) == 0:
-                    count += 1
-                    valid = False
-                    print(colored(f"Warning [{count}]: {student} has no grade for personal development item '{item}'! Please check the grader report.", "red"))
-
-        print(colored(f"Validation Pass: {valid}", "red" if not valid else "green"), "\n", colored(f"Warnings: {count}\n", "yellow") if not valid else "")
-        return valid
-
     def generate_all(self, autocorrect = True, callback = None, force = False):
-        job_count = len(self.students.index)
+        job_count = len(self.grader_report.students.index)
         
-        for i, student in enumerate(self.students.index):
+        for i, student in enumerate(self.grader_report.students.index):
             status_message = f"Generating report for {student}…"
             if callback is not None:
                 callback(i, job_count, status_message)
@@ -210,13 +48,22 @@ class Generator:
 
     def generate_for_student(self, student_name, autocorrect = True, force = False):
         # Prepare data
-        if not self.__data_valid and not force:
+        if not self.grader_report.data_valid and not force:
             print(f"[  ] Grader report incomplete. Aborting process...")
             return
         
-        str_letter_grade = str(self.get_final_grade(student_name, "Letter Grade"))
-        str_final_score = str(self.get_final_grade(student_name, "Final Score"))
+        str_letter_grade = str(self.grader_report.get_final_grade(student_name, "Letter Grade"))
+        str_final_score = str(self.grader_report.get_final_grade(student_name, "Final Score"))
 
+        comment_generator = cgen.CommentGenerator(student_name = student_name, 
+                                                  short_name = self.grader_report.get_student_info(student_name, "Short Name"),
+                                                  gender = self.grader_report.get_student_info(student_name, "Gender"),
+                                                  comment_mapping = self.grader_report.data_comment_mapping,
+                                                  student_result = student_sna,
+                                                  letter_grade = str_letter_grade,
+                                                 )
+
+        # Document processing
         document = Document()
         document = document_helper.setup_page(document, 'a4')
         
@@ -251,20 +98,20 @@ class Generator:
 
         ci_table.cell(1, 0).text = "Grade"
         ci_table.cell(1, 0).paragraphs[0].runs[0].bold = True
-        ci_table.cell(1, 1).text = self.get_course_info("Grade")
+        ci_table.cell(1, 1).text = self.grader_report.get_course_info("Grade")
         
         ci_table.cell(2, 0).text = "School Year"
         ci_table.cell(2, 0).paragraphs[0].runs[0].bold = True
-        ci_table.cell(2, 1).text = self.get_course_info("School Year")
+        ci_table.cell(2, 1).text = self.grader_report.get_course_info("School Year")
 
         ci_table.cell(0, 2).text = "Semester"
         ci_table.cell(0, 2).paragraphs[0].runs[0].bold = True
-        ci_table.cell(0, 3).merge(ci_table.cell(0, 4)).text = str(self.get_course_info("Semester"))
+        ci_table.cell(0, 3).merge(ci_table.cell(0, 4)).text = str(self.grader_report.get_course_info("Semester"))
         ci_table.cell(0, 3).paragraphs[0].alignment = WD_TABLE_ALIGNMENT.CENTER
 
         ci_table.cell(1, 2).text = "Subject"
         ci_table.cell(1, 2).paragraphs[0].runs[0].bold = True
-        ci_table.cell(1, 3).merge(ci_table.cell(1, 4)).text = self.get_course_info("Subject")
+        ci_table.cell(1, 3).merge(ci_table.cell(1, 4)).text = self.grader_report.get_course_info("Subject")
         ci_table.cell(1, 3).paragraphs[0].alignment = WD_TABLE_ALIGNMENT.CENTER
         
         ci_table.cell(2, 2).text = "Assessment"
@@ -290,7 +137,7 @@ class Generator:
         sd_table.autofit = False
         sd_table.allow_autofit = False
         sd_table.rows[0].height = Cm(1.5)
-        sd_table.cell(0, 0).text = self.get_course_info("Subject Description")
+        sd_table.cell(0, 0).text = self.grader_report.get_course_info("Subject Description")
         sd_table.cell(0, 0).paragraphs[0].alignment = WD_TABLE_ALIGNMENT.LEFT
         sd_table.cell(0, 0).width = Cm(17)
 
@@ -300,7 +147,7 @@ class Generator:
         sna_header.paragraph_format.space_before = Pt(18)
         sna_header.paragraph_format.space_after = Pt(0)
         
-        sna_table = document.add_table(rows = len(self.data_sna.columns), cols = 2)
+        sna_table = document.add_table(rows = len(self.grader_report.data_sna.columns), cols = 2)
         sna_table.style = "Table Grid"
         sna_table.alignment = WD_TABLE_ALIGNMENT.CENTER
         sna_table.autofit = False
@@ -308,8 +155,8 @@ class Generator:
 
         student_sna = {}
         
-        for i, assessment in enumerate(self.data_sna.columns):
-            student_sna[assessment] = self.get_grade_sna(student_name, assessment)
+        for i, assessment in enumerate(self.grader_report.data_sna.columns):
+            student_sna[assessment] = self.grader_report.get_grade_sna(student_name, assessment)
             sna_table.cell(i, 0).text = assessment
             sna_table.cell(i, 0).width = Cm(15)
             sna_table.cell(i, 1).text = str(student_sna[assessment])
@@ -322,7 +169,7 @@ class Generator:
         pd_header.paragraph_format.space_before = Pt(18)
         pd_header.paragraph_format.space_after = Pt(0)
         
-        pd_table = document.add_table(rows = len(self.data_pd.columns) + 1, cols = 6)
+        pd_table = document.add_table(rows = len(self.grader_report.data_pd.columns) + 1, cols = 6)
         pd_table.style = "Table Grid"
         pd_table.alignment = WD_TABLE_ALIGNMENT.CENTER
         pd_table.autofit = False
@@ -341,15 +188,15 @@ class Generator:
             pd_table_header[i].paragraphs[0].alignment = WD_TABLE_ALIGNMENT.CENTER
             pd_table_header[i].width = Cm(1) if i != 0 else Cm(12)
 
-        for i, item in enumerate(self.data_pd.columns):
+        for i, item in enumerate(self.grader_report.data_pd.columns):
             pd_table.cell(i + 1, 0).text = item
             pd_table.cell(i + 1, 0).paragraphs[0].alignment = WD_TABLE_ALIGNMENT.LEFT
             pd_table.cell(i + 1, 0).width = Cm(12)
-            pd_grade = int(self.get_grade_pd(student_name, item))
+            pd_grade = int(self.grader_report.get_grade_pd(student_name, item))
             pd_table.cell(i + 1, pd_grade).text = "✔"
             pd_table.cell(i + 1, pd_grade).paragraphs[0].alignment = WD_TABLE_ALIGNMENT.CENTER
             
-        for i in range(1, len(self.data_pd.columns) + 1):
+        for i in range(1, len(self.grader_report.data_pd.columns) + 1):
             for j in range(1, 6):
                 pd_table.cell(i, j).width = Cm(1)
 
@@ -365,13 +212,7 @@ class Generator:
         tc_table.autofit = False
         tc_table.allow_autofit = False
         tc_table.rows[0].height = Cm(2)
-        tc_table.cell(0, 0).text = cgen.CommentGenerator(student_name = student_name, 
-                                                         short_name = self.get_student_info(student_name, "Short Name"),
-                                                         gender = self.get_student_info(student_name, "Gender"),
-                                                         comment_mapping = self.data_comment_mapping,
-                                                         student_result = student_sna,
-                                                         letter_grade = str_letter_grade,
-                                                        ).generate_comment(autocorrect = autocorrect)
+        tc_table.cell(0, 0).text = comment_generator.generate_comment(autocorrect = autocorrect)
         tc_table.cell(0, 0).paragraphs[0].alignment = WD_TABLE_ALIGNMENT.LEFT
         tc_table.cell(0, 0).width = Cm(17)
 
@@ -392,7 +233,7 @@ class Generator:
             row.height = Cm(1)
 
         ak_table.cell(0, 0).paragraphs[0].add_run("Teacher:").bold = True
-        ak_table.cell(0, 0).add_paragraph(self.get_course_info("Teacher"))
+        ak_table.cell(0, 0).add_paragraph(self.grader_report.get_course_info("Teacher"))
         ak_table.cell(0, 1).text = "Signature"
         ak_table.cell(0, 2).text = "Date"
